sop/IFRAME_EMBED_OBJECT_svg.sub.html

<!DOCTYPE html>
<html>
<head>
    <title>IFRAME, EMBED, OBJECT with SVG test cases</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script type="text/javascript">

        // -------------------- DEFINITIONS -----------------------------------
        // ----- SAME-ORIGIN -----
        var test_HD_A_iframe_ED_A_r = async_test("IFRAME - HD to ED - same-origin read");
        var test_HD_A_iframe_ED_A_w = async_test("IFRAME - HD to ED - same-origin write");
        var test_HD_A_object_ED_A_r = async_test("OBJECT - HD to ED - same-origin read");
        var test_HD_A_object_ED_A_w = async_test("OBJECT - HD to ED - same-origin write");
        var test_HD_A_embed_ED_A_r = async_test("EMBED - HD to ED - same-origin read");
        var test_HD_A_embed_ED_A_w = async_test("EMBED - HD to ED - same-origin write");

        // ----- CROSS-ORIGIN -----
        var test_HD_A_iframe_ED_B_r = async_test("IFRAME - HD to ED - cross-origin read");
        var test_HD_A_iframe_ED_B_w = async_test("IFRAME - HD to ED - cross-origin write");
        var test_HD_A_object_ED_B_r = async_test("OBJECT - HD to ED - cross-origin read");
        var test_HD_A_object_ED_B_w = async_test("OBJECT - HD to ED - cross-origin write");
        var test_HD_A_embed_ED_B_r = async_test("EMBED - HD to ED - cross-origin read");
        var test_HD_A_embed_ED_B_w = async_test("EMBED - HD to ED - cross-origin write");


        // -------------------- IMPLEMENTATIONS -------------------------------
        // ----- SAME-ORIGIN -----
        function f_test_HD_A_iframe_ED_A_r() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_iframe_ED_A_r.step_func_done(function() {
                svgDoc = ee.getSVGDocument();
                var firstChildName = svgDoc.documentElement.firstElementChild.nodeName;
                assert_equals(firstChildName, "rect");
            });

            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_iframe_ED_A_w() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_iframe_ED_A_w.step_func_done(function() {
                svgDoc = ee.getSVGDocument();
                var firstChild = svgDoc.documentElement.firstElementChild;
                assert_not_equals(firstChild, null);
                svgDoc.documentElement.removeChild(firstChild);
                var firstChild = svgDoc.documentElement.firstElementChild;
                assert_equals(firstChild, null);
            });

            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_iframe_ED_A_x() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg_script.py?operation=execute&description=IFRAME - HD to ED - same-origin execute';
            document.body.appendChild(ee);
            fetch_tests_from_window(ee.contentWindow);
        }

        function f_test_HD_A_object_ED_A_r() {
            var ee = document.createElement('object');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_object_ED_A_r.step_func_done(function() {
                svgDoc = ee.getSVGDocument();
                var firstChildName = svgDoc.documentElement.firstElementChild.nodeName;
                assert_equals(firstChildName, "rect");
            });

            ee.data = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_object_ED_A_w() {
            var ee = document.createElement('object');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_object_ED_A_w.step_func_done(function() {
                svgDoc = ee.getSVGDocument();
                var firstChild = svgDoc.documentElement.firstElementChild;
                assert_not_equals(firstChild, null);
                svgDoc.documentElement.removeChild(firstChild);
                var firstChild = svgDoc.documentElement.firstElementChild;
                assert_equals(firstChild, null);
            });

            ee.data = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_embed_ED_A_r() {
            var ee = document.createElement('embed');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_embed_ED_A_r.step_func_done(function() {
                svgDoc = ee.getSVGDocument();
                var firstChildName = svgDoc.documentElement.firstElementChild.nodeName;
                assert_equals(firstChildName, "rect");
            });

            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_embed_ED_A_w() {
            var ee = document.createElement('embed');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_embed_ED_A_w.step_func_done(function() {
                svgDoc = ee.getSVGDocument();
                var firstChild = svgDoc.documentElement.firstElementChild;
                assert_not_equals(firstChild, null);
                svgDoc.documentElement.removeChild(firstChild);
                var firstChild = svgDoc.documentElement.firstElementChild;
                assert_equals(firstChild, null);
            });

            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_ED_A_iframe_HD_A_r() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg_script.py?operation=read&description=IFRAME - ED to HD - same-origin read';
            document.body.appendChild(ee);
            fetch_tests_from_window(ee.contentWindow);
        }

        function f_test_ED_A_iframe_HD_A_w() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.src = 'http://{{domains[]}}:{{ports[http][0]}}/sop/resources/svg_script.py?operation=write&description=IFRAME - ED to HD - same-origin write';
            document.body.appendChild(ee);
            fetch_tests_from_window(ee.contentWindow);
        }

        // ------------------------
        // ----- CROSS-ORIGIN -----
        // ------------------------

        function f_test_HD_A_iframe_ED_B_r() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_iframe_ED_B_r.step_func_done(function() {
                assert_equals(ee.getSVGDocument(), null);
            });

            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_iframe_ED_B_w() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_iframe_ED_B_w.step_func_done(function() {
                assert_equals(ee.getSVGDocument(), null);
            });

            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_iframe_ED_B_x() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg_script.py?operation=execute&description=IFRAME - HD to ED - cross-origin execute';
            document.body.appendChild(ee);
            fetch_tests_from_window(ee.contentWindow);
        }

        function f_test_HD_A_object_ED_B_r() {
            var ee = document.createElement('object');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_object_ED_B_r.step_func_done(function() {
                assert_equals(ee.getSVGDocument(), null);
            });

            ee.data = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_object_ED_B_w() {
            var ee = document.createElement('object');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_object_ED_B_w.step_func_done(function() {
                assert_equals(ee.getSVGDocument(), null);
            });

            ee.data = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_embed_ED_B_r() {
            var ee = document.createElement('embed');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_embed_ED_B_r.step_func_done(function() {
                assert_equals(ee.getSVGDocument(), null);
            });

            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_HD_A_embed_ED_B_w() {
            var ee = document.createElement('embed');
            ee.width = 0;
            ee.height = 0;
            ee.onload = test_HD_A_embed_ED_B_w.step_func_done(function() {
                assert_equals(ee.getSVGDocument(), null);
            });

            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg.py';
            document.body.appendChild(ee);
        }

        function f_test_ED_B_iframe_HD_A_r() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg_script.py?operation=partial read&description=IFRAME - ED to HD - cross-origin partial read';
            document.body.appendChild(ee);
            fetch_tests_from_window(ee.contentWindow);
        }

        function f_test_ED_B_iframe_HD_A_w() {
            var ee = document.createElement('iframe');
            ee.width = 0;
            ee.height = 0;
            ee.src = 'http://{{domains[www2]}}:{{ports[http][0]}}/sop/resources/svg_script.py?operation=partial write&description=IFRAME - ED to HD - cross-origin partial write';
            document.body.appendChild(ee);
            fetch_tests_from_window(ee.contentWindow);
        }

        // -------------------- FUNCTION CALLS -------------------------------
        // ----- SAME-ORIGIN -----
        f_test_HD_A_iframe_ED_A_r();
        f_test_HD_A_iframe_ED_A_w();
        f_test_HD_A_iframe_ED_A_x();
        f_test_HD_A_object_ED_A_r();
        f_test_HD_A_object_ED_A_w();
        f_test_HD_A_embed_ED_A_r();
        f_test_HD_A_embed_ED_A_w();

        f_test_ED_A_iframe_HD_A_r();
        f_test_ED_A_iframe_HD_A_w();

        // ----- CROSS-ORIGIN -----
        f_test_HD_A_iframe_ED_B_r();
        f_test_HD_A_iframe_ED_B_w();
        f_test_HD_A_iframe_ED_B_x();
        f_test_HD_A_object_ED_B_r();
        f_test_HD_A_object_ED_B_w();
        f_test_HD_A_embed_ED_B_r();
        f_test_HD_A_embed_ED_B_w();

        f_test_ED_B_iframe_HD_A_r();
        f_test_ED_B_iframe_HD_A_w();


    </script>

    <div id="log"></div>

    <div id="loadbar_IFRAME - ED to HD - same-origin write" style="display: none"><p>Not empty</p></div>

</body>
</html>
